Restrict the KV cache to int8 or bf16
int4 is now rejected by QuantConfig, the way int4 activations already were.
It was measured, not assumed: on its own it costs +23.23% of generation
accuracy, and with int4 MLP weights it takes the model below random guessing
(0.0989 against 0.25 for a four-way choice) while perplexity reports the damage
as +0.18%, because a perplexity pass never reads the cache back. int8 costs
+1.28% and halves cache traffic, so int8 to bf16 is the whole useful range.

The sweep grid drops to 16 combinations. Two existing tests asserted the old
behaviour and were updated: one parametrized int4 as a valid cache width, the
other used an int4 cache to show the cache costs nothing on disk, which an int8
cache shows equally well.

Also writes down why an int8 activation does not make anything faster, which
came up alongside this. No mode in the repo runs an int8xint8 matmul -- fake
quantize-dequantizes then multiplies in bf16, real quantizes to integers then
multiplies in fp32 for exactness, and the kernel is weight-only and reads the
activation scheme only to print a label. An int GEMM would not help either:
torch._int_mm measured 1.03-1.04x against bf16 here and needs M>16, so decode
cannot call it, and at M=1 the matmul is a GEMV whose bottleneck is streaming
the weights -- the activation is a few KB against 1.6 GB per token. That is why
the activation axis moves BPV and decode traffic by exactly 0.00.

272 tests pass.

+++ tests/test_metrics.py
@@ -137,9 +137,9 @@ def test_kv_cache_costs_nothing_on_disk_but_grows_decode_with_context():
     from llmquant.core.metrics import decode_bytes_at_context
 
     bf16 = model_metrics(TinyLlama(), QuantConfig(attn_weight="int8", kv_cache="bf16").to_modifier())
-    int4 = model_metrics(TinyLlama(), QuantConfig(attn_weight="int8", kv_cache="int4").to_modifier())
-    assert bf16["deployed_bytes"] == int4["deployed_bytes"]  # the cache is not stored
-    assert int4["kv_bytes_per_token"] < bf16["kv_bytes_per_token"]
+    int8 = model_metrics(TinyLlama(), QuantConfig(attn_weight="int8", kv_cache="int8").to_modifier())
+    assert bf16["deployed_bytes"] == int8["deployed_bytes"]  # the cache is not stored
+    assert int8["kv_bytes_per_token"] < bf16["kv_bytes_per_token"]
     # at zero context the cache costs nothing; the gap opens as context grows
-    assert decode_bytes_at_context(bf16, 0) == decode_bytes_at_context(int4, 0)
-    assert decode_bytes_at_context(int4, 8192) < decode_bytes_at_context(bf16, 8192)
+    assert decode_bytes_at_context(bf16, 0) == decode_bytes_at_context(int8, 0)
+    assert decode_bytes_at_context(int8, 8192) < decode_bytes_at_context(bf16, 8192)
